Initialize a tundraContainer object.

Initialize a tundraContainer object.

initialize(keyword, train_function = identity, predict_function = identity,
  munge_procedure = list(), default_args = list(), internal = list())

Arguments

keyword
character. The name of the classifier; for example, "lm" or "knn".

train_function
function. The function used to train the model. Its first argument will be a data.frame, and the second argument a list of additional parameters used for training the model.

predict_function
function. The function used to predict on new datasets. Its first argument will be a data.frame, the dataset to predict on, and its second (optional) argument will be additional parameters used for prediction output (such as whether to return a probabilistic or absolute value).

munge_procedure
list. A list of trained mungepieces to apply to data sets during prediction.

default_args
list. A list of default arguments to provide to the second argument of the train_function. The additional arguments provided to the tundraContainer's train method will be merged on top of these defaults.

internal
list. Internal metadata that should accompany the model. Usually this is domain/organization specific, and can include things such as a list of primary keys used for training the model, identifiers or names of data sources used for training the model, etc. It is a playground entirely under your control, and can be used by other packages or a production server hosting the model to achieve additional behavior.